Struct http_auth_basic::Credentials

pub struct Credentials {
    pub user_id: String,
    pub password: String,
}

A struct to represent the user_id and password fields from an Authorization Basic header value

Fields
user_id: Stringpassword: String
Implementations

impl Credentials

pub fn new(user_id: &str, password: &str) -> Self

Create a new Credentials instance this is equivalent to writing:

use http_auth_basic::Credentials;

let credentials = Credentials {
    user_id: String::from("Foo"),
    password: String::from("Bar"),
};

pub fn decode(auth_header_value: String) -> Result<Self, AuthBasicError>

Create a Credentials instance from a base64 String which must encode user credentials as username:password

pub fn encode(&self) -> String

Encode a Credentials instance into a base64 String

pub fn from_header(auth_header: String) -> Result<Credentials, AuthBasicError>

Creates a Credentials instance from an HTTP Authorization header which schema is a valid Basic HTTP Authorization Schema.

pub fn as_http_header(&self) -> String

Creates a HTTP Authorization header value for the basic schema from the Credentials instance